🎨 style: LevelSection 컴포넌트 분리 #28
